Improve/ez 143 rename module backtest (#63)
* feat: rename backtest to backtesting

* doc: update doc
