+++ ezyquant/__init__.py
@@ -1,4 +1,4 @@
-from .backtest import backtest
+from .backtesting import backtest
 from .connect import connect_sqlite
 from .creator import SETSignalCreator
 from .reader import SETDataReader
