+++ ezyquant/backtesting/__init__.py
@@ -1,5 +1,5 @@
 from .account import SETAccount
-from .backtest import backtest
+from .backtesting import backtest
 from .context import Context
 from .position import SETPosition
 from .trade import SETTrade
